+++ airflow/contrib/operators/emr_terminate_job_flow_operator.py
@@ -16,9 +16,9 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from airflow.contrib.hooks.emr_hook import EmrHook
 from airflow.exceptions import AirflowException
 from airflow.models import BaseOperator
+from airflow.providers.amazon.aws.hooks.emr import EmrHook
 from airflow.utils.decorators import apply_defaults
